dockerize pattern count rename folder and READMEs
