--- Count_Patern/package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "express": "^4.21.2",
-    "md5": "^2.3.0"
-  }
-}
