--- Count_Patern/public/user.html
@@ -1,41 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>User</title>
-    <link rel="stylesheet" href="style.css">
-</head>
-<body>
-<div class="container">
-    <h2>User Area</h2>
-    <form onsubmit="compter(event)">
-        <label for="texte">Text:</label>
-        <textarea id="texte" rows="4"></textarea>
-        <label for="pattern">Word or Pattern to search:</label>
-        <input type="text" id="pattern" />
-        <button type="submit">Count</button>
-    </form>
-    <div class="result" id="result"></div>
-    <button class="logout" onclick="logout()">Logout</button>
-</div>
-<script>
-    async function compter(event) {
-        event.preventDefault();
-        const text = document.getElementById("texte").value;
-        const pattern = document.getElementById("pattern").value;
-        const response = await fetch("/compter", {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({ text, pattern }),
-        });
-        const data = await response.json();
-        document.getElementById("result").textContent = data.result || "Error";
-    }
-
-    function logout() {
-        sessionStorage.removeItem("role");
-        window.location.href = "login.html";
-    }
-</script>
-</body>
-</html>
