+++ pattern_counter/.dockerignore
@@ -1,0 +1,8 @@
+.git
+.gitignore
+README.md
+node_modules
+dist
+.env
+Dockerfile
+docker-compose.yaml
