+++ pattern_counter/package.json
@@ -1,0 +1,13 @@
+{
+  "dependencies": {
+    "dotenv": "^16.4.7",
+    "express": "^4.21.2",
+    "md5": "^2.3.0"
+  },
+  "devDependencies": {
+    "prettier": "^3.4.2"
+  },
+  "scripts": {
+    "start": "node server.js"
+  }
+}
