+++ pattern_counter/public/db.csv
