+++ pattern_counter/public/index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <title>Pattern Counter</title>
+    <link rel="stylesheet" href="style.css">
+</head>
+
+<body>
+    <div class="container">
+        <h1>Pattern Counter</h1>
+        <p>Easily count the number of times a pattern appears in a text</p>
+        <form action="login.html" method="get">
+            <button type="submit">Login</button>
+        </form>
+    </div>
+</body>
+
+</html>
