+++ pattern_counter/public/login.html
@@ -1,0 +1,58 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <title>Login</title>
+  <link rel="stylesheet" href="style.css">
+</head>
+
+<body>
+  <div class="container">
+    <h1>Login</h1>
+    <form id="loginForm">
+      <label for="username">Username:</label>
+      <input type="text" id="username" required />
+      <label for="password">Password:</label>
+      <input type="password" id="password" required />
+      <button type="submit">Login</button>
+    </form>
+    <div id="message"></div>
+  </div>
+  <script>
+    document.getElementById("loginForm").onsubmit = async (event) => {
+      event.preventDefault();
+      const username = document.getElementById("username").value;
+      const password = document.getElementById("password").value;
+
+      const response = await fetch("/login", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ username, password }),
+      });
+
+      const result = await response.json();
+      if (result.success) {
+        if (result.role === "admin") {
+          const flagResponse = await fetch("/get-admin-flag", {
+            method: "POST",
+            headers: { "Content-Type": "application/json" },
+            body: JSON.stringify({ sessionId: result.sessionId }),
+          });
+          const flagResult = await flagResponse.json();
+          if (flagResult.success) {
+            document.getElementById("message").textContent = `${flagResult.flag}`;
+          } else {
+            document.getElementById("message").textContent = "Could not retrieve the flag.";
+          }
+        } else if (result.role === "user") {
+          window.location.href = "/user.html";
+        }
+      } else {
+        document.getElementById("message").textContent = "Invalid username or password. Please try again. The password can only contain numbers and have 6 characters maximum.";
+      }
+    };
+  </script>
+</body>
+
+</html>
