+++ pattern_counter/public/secret.html
@@ -1,14 +1,17 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <title>Admin Secret Area</title>
     <link rel="stylesheet" href="style.css">
 </head>
+
 <body>
-<div class="container">
-    <h2>Admin Secret Area</h2>
-    <p>flag{I_thought_that_I_hid_better}</p>
-</div>
+    <div class="container">
+        <h2>Admin Secret Area</h2>
+        <p>flag{I_thought_that_I_hid_better}</p>
+    </div>
 </body>
+
 </html>
