+++ pattern_counter/public/secret_page_that_definitely_doesnt_have_a_flag_hide_inside.html
@@ -1,13 +1,16 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
   <meta charset="UTF-8">
   <title>Secret Page</title>
   <link rel="stylesheet" href="style.css">
 </head>
+
 <body>
-<div class="container">
-  <p>flag{this_is_definitely_not_a_flag__please_ignore_it}</p>
-</div>
+  <div class="container">
+    <p>flag{this_is_definitely_not_a_flag__please_ignore_it}</p>
+  </div>
 </body>
+
 </html>
