+++ pattern_counter/public/style.css
