+++ pattern_counter/solution/solve.py
